Moves categories helper into helper folder
