+++ convert.js
@@ -8,7 +8,7 @@ let _ = require('lodash');
 const yargs = require('yargs');
 
 // Project file imports
-let categories = require('./categories');
+let categories = require('./helpers/categories');
 let fileOpsHelper = require('./helpers/fileops');
 let taxHelper = require('./helpers/taxhelper');
 let vmsHelper = require('./helpers/vmshelper');
