+++ helpers/categories.js
